fix: config.workspaceDir should be honored when cliFlags.input is given
Before this fix, when input file is specified in command line arguments, the `entryContextDir` and `workspaceDir` are set to the location of the input file, even if vivliostyle config file exists. This causes the following bug:

- #402

After this fix, when the config file exists the `entryContextDir` and `workspaceDir` are given from the properties in the config file or set to the location of the config file.

fix #402.